requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.9.0
django-filter==25.2
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.2
pillow==11.3.0
psycopg==3.2.10
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
gunicorn==21.2.0
whitenoise==6.7.0           
drf-spectacular==0.27.1
africastalking==1.2.5